importer: Dont clean netpath imports
When getting the directory of an import and appending an import to an
element of the search path, the `Importer` uses `path.Dir()` and
`path.Join()`. However these functions clean the path returned which
includes collapsing a run of consecutive slashes into a single slash.
This removes the leading double slash we use to indicate a netpath.

Handle this by adding back the extra slash if it was removed.

This addresses #6[1] (Importer fails to import netpaths). It might make
sense to implement a `netpath` package that solves this problem properly
and defines the semantics of netpaths (i.e. what is the dir of
`//example.com`?), but for now we'll make do with this point solution.

Issue: [1] https://github.com/foxygoat/jsonnext/issues/6