build: Clean up Makefile
Rename `lint-local` to `lint-with-local`. This accidentally got changed
in a Makefile cleanup.

Clean up the Makefile in other ways, such as removing blank lines, using
`$(O)` instead of `$O` as we should probably do for all variable
expansions as it makes it a little clearer to scan.

This should match the other foxygoat project makefiles now, as we
converge on a common format.