importer: Do not construct relative file:// URLs
`file://` URLs cannot represent relative paths. The importer assumes
that they can by using a double-slash instead of a triple-slash. This is
wrong.

Force the path to be absolute when constructing a `file://` URL.

Drop coverage from 100 to 96.5 because it's really hard to make
`filepath.Abs()` return an error.

Issue: https://github.com/foxygoat/jsonnext/issues/12