build: Make lint-with-docker target in Makefile faster
Running golangci-lint in docker goes fastest when you maintain the cache
across runs, both the Go cache and the lint cache. This is in
`~/.cache` on the host filesystem and `/root/.cache` in the container (I
guess lint runs as root, naughty). Map this volume when running lint via
docker.

Docker still adds about 1.1s of overhead running lint. A fully cached
lint run outside docker takes about 250ms. In docker it takes 1350ms.
The time of lint itself in docker is 250ms, so the overhead is the
docker setup for the container. Somewhat disappointing we can't achieve
the same speeds with docker than without.