ci: Split version bump into separate workflow
The version bump action was at the and of CI conditional on the branch
being master. This meant it would wait for CI to finish but not run on
branches. However it still shows up on a PR as a skipped step which is
annoying.

The alternative is to make the version bump a separate workflow, which
is what this commit does. But GitHub Actions does not allow you to make
one workflow depend on another, so this will run in parallel with CI on
master. This means if CI fails, we still bump the version. This should
not happen as we require PRs to have passing CI and also require them to
be up-to-date WRT master, so it would be odd if CI passes on a PR but
fails on master when merged.

This removes the "noise" on each PR of the skipped version bump step.