importer: Stop using file:/// URLs for local imports
The importer was converting local imports to `file:///` URLs and
importing them using a `file` protocol handler in the `URLFetcher`
(`http.Client`). This has two problems:

1. If you try to import a directory, you get an HTML directory listing,
   instead of an error.
2. If you try to import a stream (`/dev/stdin` for instance), it fails
   as http.FileServer wants seekable files.

It also was not well documented that if you provide your own fetcher, it
needs to handle `file:///` URLs.

Dump this and just read the files directly.

This seems to make the tests cleaner - they have been pretty much
re-written for this and use a `testdata` directory to drive them. They
are a little more blackbox than unit tests which is kind of nice as the
unit tests assumed too much about the internal behaviour. Testing the
importer from the outside should give more confidence it does what it is
meant to - something that was lost with the initial bugs but 100% test
coverage, due to invalid assumptions.