jx: Add jx command to run jsonnet code
Add a jx command that uses `jsonnext.Importer` to import jsonnet files,
and run the initial file to be evaluated through the importer too. This
means the jsonnet code to be evaluated does not need to be local:

    jx -J //github.com/yugui/jsonnetunit/raw/master jsonnetunit/test/std_matchers_test.jsonnet

Will eval the `jsonnetunit/test/std_matchers_test.jsonnet` file from
github.com/yugui/jsonnetunit, which imports `jsonnetunit/test.libsonnet`
from that same repository.

There is currently no way to pass in top-level args or extvars - that
will come shortly when the flags parsing for them have been added.

It is expected that `jx` will expand to use most of the features present
in libraries that get added to this repository to serve as an example
and to also provide a useful command.