config: Add Config struct to configure jsonnet VM
Add a Config struct that can be used to collect combined jsonnet VM
configuration to apply to a VM and/or Importer.

It is mostly centered around VM variables (ext vars, top-level args) for
now, with a types to represent each type of variable and a map for them,
as well as parsing for setting variables from strings (such as the
command line).

This will be used in later commits with CLI flag parsing helpers, but
could also be extended to read from a config file. The base intent is to
allow an application to take its configuration from whatever sources it
uses to populate the Config struct and to configure the VM in one hit.