kong: Parse CLI into Config using kong
Add a kong package to parse CLI args into the Config struct using kong,
and update the jx command to use it.

Augment the jsonnext.Config struct with struct tags for kong where we
can, but that is not sufficient for the var maps as each has multiple
flags that populate them, which kong cannot do directly. So this package
provides an auxiliary struct for all the command line flags we need to
parse using a custom type with a kong Decode method.

Add a conformance test suite to test that the kong CLI parser correctly
parses command line flags into the Config struct. The conformance test
suite is set up so it can be used by any CLI parsing implementation to
verify its correctness.

Provide an example (which serves to give almost complete test coverage)
on how to use it.