flags: Add Go flag-compatible helpers for Config
Add functions for defining flags for the Go flag package to parse the
CLI into a Config struct. These follow a similar pattern to the function
naming in the Go `flag` package with `Foo` and `FooVar` where `Foo`
returns a Foo bound to the flag and `FooVar` binds a flag to the given
`Foo`.

Use the conformance test suite to test most of the flags functionality,
adding only specific tests for code that is not covered by the
conformance test suite.

Split the `jx` main into two builds so we can actually run the flags
stuff for testing purposes. `jx` primarily uses kong to parse the CLI,
but we want to actually make use of these flag functions, so we build
jx with the "flag" tag to produce out/jx-flag.