lib: Add a jsonnet utilities library
Add a library of jsonnet code for general purpose manipulation of
jsonnet values.

This is the first component, `values.jsonnet` that contains some general
purpose functions operating on any/all types.

Extend the Makefile to run the tests on this library of jsonnet code,
using the `jx` binary to so do, so that jsonnetunit can be imported via
the network. The test harness can be executed directly if the `jx`
binary is on the PATH.

Also make some small updates to the Makefile to simplify it - there is
no need to have make find all the directories not starting with an
underscore - the Go toolchain does that fine with `./cmd/...`.

Subsequent commits will add libraries for other specific types.