lib: Fix filename import in top-level jx.jsonnet
The top-level `jx.jsonnet` was importing `value.jsonnet` as
`value.libsonnet` as I made a recent decision to stop using the
`.libsonnet` extension, but I forgot to update that one reference.

Add a test to `jx_test.jsonnet` to test that all the imports work, and
that everything imported is hidden.