Update all dependencies
Update all Go dependencies with:

    go get -u all
    go mod tidy

This is mostly to get go-jsonnet v0.17.0 so I can get some new library
functions for my tests, but may as well keep up-to-date.

This has caused some new lint warnings as EvaluateSnippet has become
deprecated. Update to use the non-deprecated EvaluateAnonymousSnippet
instead. Also fix a silly linter error that can't read godocs properly
(I think it thought the comment line starting with `// package: ` was a
compiler/tool directive and not just part of the comment, so it ignored
that and the period that was at the end, and then complained about a
missing period. Changing the `:` to ` -` has removed that warning.