build: Use latest foxygoat like Makefile snippet
Update the Go linting `Makefile` snippet to the lastest foxygoat
standand (from foxtrot), and update it for Go 1.16 which allows `go
install` with a version from a module.

This no longer uses docker for a specific version as it is too slow.
Instead we build the version we need if we don't have it.