flag: Add --max-{stack,trace} flags
Add the --max-stack and --max-trace flags to set the maximum jsonnet VM
stack size (in frames) and the maximum number of frames printed in error
output. The defaults of 500 and 20 respectively come from the jsonnet
source. It would be nice to not set these values on the VM if the flags
are not present, but it's also nice to know from the cli help what the
defaults are. I went with the latter.

The flags do not use the `-s`/`-t` short flags used by the jsonnet
binary. These options are rather esoteric and I don't feel they deserve
a short flag.