tools: Hermitise repo and pin versions required
Hermitise this repository as the version of golangci-lint used is not
compatible with more recent Go releases so we will either need to pin to
an older version of Go or upgrade golangci-lint. The former is tricky to
do simply, and the latter is painful with the churn in golangci-lint.

The alternative is to use hermit and pin the versions currently
required. This is what is done here:

    hermit init --no-git
    hermit install go-1.13.5
    hermit install golangci-lint-1.37.0

Remove the custom version login for golangci-lint from the Makefile, and
add the hermit auto-activation. Note that the only 1.13 version of Go in
hermit is 1.13.5, while the latest point release is 1.13.15. I chose
1.13 as that is what is in the `go.mod` file.

Subsequent commits will bump these versions and adjust the code to work
with newer golangci-lint errors.

Update GitHub CI to rely on hermit for installing the right versions
instead of using specific actions. Also run on all PRs, not just those
to be merged to master. This allows CI to run for stacked PRs.